feat(category): add send and recieve prop
