+++ src/assets/scss/searchfilter.scss
@@ -97,7 +97,7 @@
                         }
                     }
                 }
-                .category {
+                .category-tags {
                     .tags {
                         display: flex;
                         flex-direction: row;
@@ -114,6 +114,9 @@
                             div {
                                 margin: auto 5px auto 0;
                             }
+                        }
+                        &.show .category {
+                            margin: 8px 10px 0 0;
                         }
                     }
                 }
